Do not trigger ScrollButton event unless clicked
