+++ panel/models/column.ts
@@ -1,6 +1,14 @@
-import {Column as BkColumn, ColumnView as BkColumnView} from "@bokehjs/models/layouts/column"
+import {ModelEvent} from "@bokehjs/core/bokeh_events"
 import {div} from "@bokehjs/core/dom"
 import type * as p from "@bokehjs/core/properties"
+import type {EventCallback} from "@bokehjs/model"
+import {Column as BkColumn, ColumnView as BkColumnView} from "@bokehjs/models/layouts/column"
+
+export class ScrollButtonClick extends ModelEvent {
+  static {
+    this.prototype.event_name = "scroll_button_click"
+  }
+}
 
 export class ColumnView extends BkColumnView {
   declare model: Column
@@ -70,6 +78,7 @@ export class ColumnView extends BkColumnView {
     })
     this.scroll_down_button_el.addEventListener("click", () => {
       this.scroll_to_latest()
+      this.model.trigger_event(new ScrollButtonClick())
     })
   }
 
@@ -117,5 +126,9 @@ export class Column extends BkColumn {
       scroll_button_threshold: [Int, 0],
       view_latest: [Bool, false],
     }))
+  }
+
+  on_click(callback: EventCallback<ScrollButtonClick>): void {
+    this.on_event(ScrollButtonClick, callback)
   }
 }
